MineralScavenger
An iOS game created with SpriteBuilder and Cocos2D. Tilting the mobile
device player can move a game character freely on screen to collect the
falling object from the sky.  This game relies on the gyro and
CMMotionManager to determine the character movement and
CCPhysicsCollisionDelegate to handle physics related collisions. It
also includes Game Center leader board and inApp purchase integration.